Pendiente integrar vistas

+++ controller/CRUD/ventasCRUD.py
@@ -57,8 +57,9 @@ class VentasCRUD:
                 self.db = self.connectSession()
             sql = f'''
                 INSERT INTO VENTAS (IDVENTA, IDSUCURSAL, IDESTADO, IDEMPLEADO, NITCLIENTE, IDFORMADEPAGO, SERIE, FACTURA, FECHAHORA, ANIOVENTA, MESVENTA, DOCUMENTOPAGO)
-                VALUES ({id}, {venta.idSucursal}, {venta.idEstado}, {venta.idEmpleado}, '{venta.nitCliente}', {venta.idFormaDePago}, '{venta.serie}', '{venta.factura}', '{venta.fechahora}', {venta.anioventa}, {venta.mesventa}, '{venta.documentoPago}')
+                VALUES ({id}, {venta.idSucursal}, {venta.idEstado}, {venta.idEmpleado}, '{venta.nitCliente}', {venta.idFormaDePago}, '{venta.serie}', '{venta.factura}', TO_DATE('{venta.fechahora}', 'yyyy-mm-dd hh24:mi:ss'), {venta.anioventa}, {venta.mesventa}, '{venta.documentoPago}')
             '''
+            print (sql)
             self.db.execute(sql)
             self.db.connection.commit()
         except Error as error:
@@ -74,13 +75,18 @@ class VentasCRUD:
     def deleteVenta(self, venta: Venta):
         try:
             data = False
-            if(self.db._verify_open):            
-                self.db = self.connectSession()
-            sql = f'''
-                DELETE FROM VENTAS WHERE IDVENTA = {venta.idVenta}
-            '''
-            self.db.execute(sql)
-            self.db.connection.commit()
+            detalleVenta=DetallesVentasCRUD()
+            data=detalleVenta.deleteDetalleVentaByIdVenta(venta.idVenta)
+            if data:
+                data=False
+                if(self.db._verify_open):            
+                    self.db = self.connectSession()
+                sql = f'''
+                    DELETE FROM VENTAS WHERE IDVENTA = {venta.idVenta}
+                '''
+                print (sql)
+                self.db.execute(sql)
+                self.db.connection.commit()
         except Error as error:
             print(error)
         except Exception as exception:
@@ -105,35 +111,13 @@ class VentasCRUD:
                     IDFORMADEPAGO = {venta.idFormaDePago}, 
                     SERIE = '{venta.serie}', 
                     FACTURA = '{venta.factura}', 
-                    FECHAHORA = '{venta.fechahora}', 
+                    FECHAHORA = TO_DATE('{venta.fechahora}', 'yyyy-mm-dd hh24:mi:ss'), 
                     ANIOVENTA = {venta.anioventa}, 
                     MESVENTA = {venta.mesventa}, 
                     DOCUMENTOPAGO = '{venta.documentoPago}'
                 WHERE IDVENTA = {venta.idVenta}
             '''
-            self.db.execute(sql)
-            self.db.connection.commit()
-        except Error as error:
-            print(error)
-        except Exception as exception:
-            print(exception)
-        else:            
-            self.db.close()
-            data = True
-        finally:
-            return data
-        
-    def updateTotalVenta(self, venta: Venta):
-        try:
-            detalleVenta=DetallesVentasCRUD()
-            data = False
-            if(self.db._verify_open):            
-                self.db = self.connectSession()
-            sql = f'''
-                UPDATE VENTAS 
-                SET TOTALVENTA = '{detalleVenta.getTotalVentaByID(Venta.idVenta)}'
-                WHERE IDVENTA = {venta.idVenta}
-            '''
+            print (sql)
             self.db.execute(sql)
             self.db.connection.commit()
         except Error as error:
